fix(tmux): Fix finally vertical split in tmux
While splitting vertically in tmux, I had the issue that the new split
was spawned in my ~ directory, while I would have liked it to be spawned
in the directory I was working.

The issue was actually trivial, despite spending far too much time on
it: I had defined the _same_ keybinding for a vertical split _not in my
current directory_ later in the file.

So, yeah, by removing those other keybinds I don't want, it fixed my
issue…